Changed namepace according to standard. Fixed test that were not running.
