+++ README.md
@@ -20,7 +20,8 @@ In most cases functions return JSON that is decoded using [`json_decode`](http:/
 *See PHPDoc comments to find out more about data returned by each function.*
 
 ## Usage example
-    use \Locomotive\Locomotive;
+
+    use Tsukanov\SteamLocomotive\Locomotive;
     $steam = new Locomotive(YOUR_STEAM_API_KEY);
 	
 	// Getting information about heroes in Dota 2
