--- src/Locomotive/Core/Tool.php
@@ -1,17 +1,0 @@
-<?php
-namespace Locomotive\Core;
-
-use Guzzle\Http\Client;
-
-class Tool
-{
-
-    public function getContent($url)
-    {
-        $client = new Client($url);
-        $request = $client->get();
-        $response = $request->send();
-        return $response->getBody(true);
-    }
-
-}
