+++ src/Tsukanov/SteamLocomotive/Core/Tool.php
@@ -1,0 +1,17 @@
+<?php
+namespace Tsukanov\SteamLocomotive\Core;
+
+use Guzzle\Http\Client;
+
+class Tool
+{
+
+    public function getContent($url)
+    {
+        $client = new Client($url);
+        $request = $client->get();
+        $response = $request->send();
+        return $response->getBody(true);
+    }
+
+}
