--- tests/Locomotive/Tests/Core/Interfaces/TestStuff.php
@@ -1,15 +1,0 @@
-<?php
-namespace Locomotive\Tests\Core\Interfaces;
-
-use Locomotive\Locomotive;
-
-class TestStuff extends \PHPUnit_Framework_TestCase
-{
-
-    public function testApiList()
-    {
-        $steam = new Locomotive("poop");
-        $list = $steam->ISteamWebAPIUtil->GetSupportedAPIList();
-    }
-
-}
