--- tests/Locomotive/Tests/Core/Tools/TestApp.php
@@ -1,18 +1,0 @@
-<?php
-namespace Locomotive\Tests\Core\Tools;
-
-use Locomotive\Core\Tools\App;
-
-class TestApp extends \PHPUnit_Framework_TestCase
-{
-
-    public function testGetAppLogoURL()
-    {
-        $app_id = 10;
-        $expected = 'http://cdn.steampowered.com/v/gfx/apps/' . $app_id . '/header.jpg';
-        $app_tools = new App();
-        $actual = $app_tools->getAppLogoURL($app_id);
-        $this->assertEquals($expected, $actual);
-    }
-
-}
