+++ tests/bootstrap.php
@@ -1,5 +1,5 @@
 <?php
 require dirname(__DIR__) . '/vendor/autoload.php';
 
-ini_set('error_reporting', E_ALL);
+ini_set('error_reporting', E_ALL | E_STRICT);
 ini_set('display_errors', 1);
